requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
colorlog==6.7.0
discord==2.2.3
discord-typings==0.6.0
discord.py==2.2.3
dnspython==2.3.0
emoji==2.4.0
frozenlist==1.3.3
idna==3.4
install==1.3.5
multidict==6.0.4
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2023.3
tomli==2.0.1
typing_extensions==4.6.2
yarl==1.9.2